Refactor poll item factory to make it testable.
